Making changes to not crap out in node
NodeJS implements a variant of CommonJS module system, which we adapt to
using the amdefine module. On the web, this is unnecessary as RequireJS
will provide the define() function.

Specifically, jquery has been removed as a library dependency from nbd
code itself, although nbd does expect elements that the View works with
to follow jQuery API.

The build step has also changed to only provide the `nbd` global
variable when using the built single-file. All other environments will
not have it.