View/Entity: clarifying render() behavior
render() now when called will almost absolutely render except in the
case when ther is both an existing $view and a new $parent. In that
case, $view will simply be appended to $parent.

We remove the case of pretemplated $view as a string because it is not a
generic usecase.