util/curry: reworked curry to use context
Instead of curry being a function that takes a single function as the
argument, use the context as the function to be curried, in the way of
util/construct.

This way to curry with predefined args:

    curry.call(fn, arg1, ...)

The old behavior of creating a currying function can still be gotten by:

    curried = curry.bind(fn)

This leads to less confusion about the number of functions involved.