trait/pubsub: replacing trigger mechanism
Mechanism relied on array filtering on the trigger side to remove called
entries marked as `once`. This has been replaced with an in-place
removal of the entry. Additionally, the event attaching side concats to
a new array instead of pushes to the existing one in order to avoid
mid-callback modifications to the array.